// app/utils/Interfaces.ts
export interface JobItem {
  id: number;
  documentId: string;
  title: string;
  slug: string;
  location: string;
  experience: string;
  description: string;
  responsibilities: string | null;
  requirements: string;
  isOpen: boolean;
  skills: string[];
  Seo?: SeoMeta;

  createdAt: string; // ISO date string
  updatedAt: string; // ISO date string
  publishedAt: string; // ISO date string
  Type: string | null;
}

export interface Meta {
  pagination: Pagination;
}

export interface Pagination {
  page: number;
  pageSize: number;
  pageCount: number;
  total: number;
}

export interface SeoMeta {
  id: number;
  metaTitle: string;
  metaDescription: string;
  keywords: string;
  metaRobots: string | null;
  metaViewport: string | null;
  canonicalURL: string | null;
  structuredData: string | null;
}
// ...
export type FieldType =
  | 'text'
  | 'email'
  | 'number'
  | 'tel'
  | 'url'
  | 'password'
  | 'textarea'
  | 'checkbox'
  | 'select'
  | 'radio'
  | 'file'
  | 'date';

export interface DynamicField {
  id?: number;
  text: string;
  label: string;
  type: FieldType;
  required?: boolean;
  options?: string[];
}